renamed record.js store to recordData.js
