+++ tests/unit/components/Curators/MaintenanceRequests.spec.js
@@ -3,7 +3,7 @@ import MaintenanceRequest from "@/components/Curators/MaintenanceRequests.vue"
 import fakeData from "@/../tests/fixtures/curationDashboardMaintReqData.json"
 import Client from "@/components/Client/RESTClient.js"
 import Vuex from "vuex"
-import recordStore from "@/store/record.js"
+import recordStore from "@/store/recordData.js"
 import sinon from "sinon"
 import usersStore from "@/store/users"
 import VueRouter from "vue-router"
