+++ tests/unit/components/Editor/AdditionalInformation/EditAdditionalInfo.spec.js
@@ -2,7 +2,7 @@ import { createLocalVue, shallowMount } from "@vue/test-utils";
 import Vuex from "vuex";
 import Vuetify from "vuetify"
 import EditAdditionalInfo from "@/components/Editor/AdditionalInformation/EditAdditionalInfo.vue"
-import recordStore from "@/store/record.js";
+import recordStore from "@/store/recordData.js";
 import userStore from "@/store/users.js";
 import editorStore from "@/store/editor.js"
 import VueRouter from "vue-router";
