+++ tests/unit/components/Editor/DataAccess/EditDataAccess.spec.js
@@ -5,7 +5,7 @@ import VueRouter from "vue-router"
 import GraphClient from "@/components/GraphClient/GraphClient.js"
 import RestClient from "@/components/Client/RESTClient.js"
 import DataAccess from "@/components/Editor/DataAccess/EditDataAccess.vue"
-import recordStore from "@/store/record.js"
+import recordStore from "@/store/recordData.js"
 import editorStore from "@/store/editor.js"
 import userStore from "@/store/users.js"
 const sinon = require("sinon");
