+++ tests/unit/components/Editor/EditPublications.spec.js
@@ -3,7 +3,7 @@ import Vuex from "vuex";
 import Vuetify from "vuetify"
 import VueRouter from "vue-router"
 import editPublications from "@/components/Editor/EditPublications.vue"
-import recordStore from "@/store/record.js";
+import recordStore from "@/store/recordData.js";
 import editorStore from "@/store/editor.js";
 import userStore from "@/store/users.js";
 import GraphClient from "@/components/GraphClient/GraphClient.js";
