+++ tests/unit/components/Editor/EditRelations.spec.js
@@ -3,7 +3,7 @@ import Vuex from "vuex";
 import Vuetify from "vuetify"
 import VueRouter from "vue-router"
 import editRelations from "@/components/Editor/EditRelationships.vue"
-import recordStore from "@/store/record.js";
+import recordStore from "@/store/recordData.js";
 import editorStore from "@/store/editor.js";
 import userStore from "@/store/users.js";
 import RestClient from "@/components/Client/RESTClient.js"
