+++ tests/unit/components/Editor/GeneralInformation/EditTags.spec.js
@@ -2,7 +2,7 @@ import { createLocalVue, shallowMount } from "@vue/test-utils"
 import Vuex from "vuex"
 import Vuetify from "vuetify"
 import EditTags from "@/components/Editor/GeneralInformation/EditTags.vue"
-import recordStore from "@/store/record.js";
+import recordStore from "@/store/recordData.js";
 import editorStore from "@/store/editor.js";
 import GraphClient from "@/components/GraphClient/GraphClient.js"
 const sinon = require("sinon");
