+++ tests/unit/components/Editor/GeneralInformation/GeneralInformation.spec.js
@@ -4,7 +4,7 @@ import Vuex from "vuex"
 import Vuetify from "vuetify"
 import VueRouter from "vue-router"
 import GeneralInfo from "@/components/Editor/GeneralInformation/GeneralInformation.vue"
-import recordStore from "@/store/record.js"
+import recordStore from "@/store/recordData.js"
 import editorStore from "@/store/editor.js"
 import userStore from "@/store/users.js"
 import GraphClient from "@/components/GraphClient/GraphClient.js"
