+++ tests/unit/components/Editor/Organisations/LinkOverlay.spec.js
@@ -3,7 +3,7 @@ import Vuex from "vuex"
 import Vuetify from "vuetify"
 import RestClient from "@/components/Client/RESTClient.js"
 import LinkOverlay from "@/components/Editor/Organisations/LinkOverlay.vue"
-import recordStore from "@/store/record.js"
+import recordStore from "@/store/recordData.js"
 import editorStore from "@/store/editor.js"
 import userStore from "@/store/users.js"
 const sinon = require("sinon");
