+++ tests/unit/components/Records/Record/DataConditions.spec.js
@@ -1,6 +1,6 @@
 import {shallowMount, createLocalVue} from "@vue/test-utils";
 import Vuex from "vuex";
-import Record from "@/store/record.js"
+import Record from "@/store/recordData.js"
 import DataConditions from "@/components/Records/Record/DataConditions.vue"
 import Vuetify from "vuetify"
