+++ tests/unit/components/Records/Record/GeneralInfo/Keywords.spec.js
@@ -1,6 +1,6 @@
 import { shallowMount, createLocalVue } from "@vue/test-utils";
 import Vuex from "vuex";
-import Record from "@/store/record.js"
+import Record from "@/store/recordData.js"
 import Keywords from "@/components/Records/Record/GeneralInfo/Keywords.vue"
 
 const localVue = createLocalVue();
