+++ tests/unit/components/Records/Record/GeneralInfo/Maintainers.spec.js
@@ -1,6 +1,6 @@
 import { shallowMount, createLocalVue } from "@vue/test-utils";
 import Vuex from "vuex";
-import Record from "@/store/record.js"
+import Record from "@/store/recordData.js"
 import Maintainers from "@/components/Records/Record/GeneralInfo/Maintainers.vue"
 import Vuetify from "vuetify"
