+++ tests/unit/components/Records/Record/GeneralInfo/Registry.spec.js
@@ -1,6 +1,6 @@
 import { shallowMount, createLocalVue } from "@vue/test-utils";
 import Vuex from "vuex";
-import Record from "@/store/record.js"
+import Record from "@/store/recordData.js"
 import Registry from "@/components/Records/Record/GeneralInfo/Registry.vue"
 import Vuetify from "vuetify"
