+++ tests/unit/components/Records/Record/Publications.spec.js
@@ -1,6 +1,6 @@
 import { shallowMount, createLocalVue } from "@vue/test-utils";
 import Vuex from "vuex";
-import Record from "@/store/record.js"
+import Record from "@/store/recordData.js"
 import Publications from "@/components/Records/Record/Publications.vue"
 
 const localVue = createLocalVue();
