+++ tests/unit/components/Records/Record/Support.spec.js
@@ -1,6 +1,6 @@
 import { shallowMount, createLocalVue } from "@vue/test-utils";
 import Vuex from "vuex";
-import Record from "@/store/record.js"
+import Record from "@/store/recordData.js"
 import Support from "@/components/Records/Record/Support.vue"
 import Vuetify from "vuetify"
