+++ tests/unit/components/Records/Record/Tools.spec.js
@@ -1,6 +1,6 @@
 import {shallowMount, createLocalVue} from "@vue/test-utils";
 import Vuex from "vuex";
-import Record from "@/store/record.js"
+import Record from "@/store/recordData.js"
 import Tools from "@/components/Records/Record/Tools.vue"
 import Vuetify from "vuetify"
