+++ tests/unit/views/Records/Record.spec.js
@@ -6,7 +6,7 @@ import VueRouter from "vue-router";
 import Record from "@/views/Records/Record.vue";
 import GraphClient from "@/components/GraphClient/GraphClient.js";
 import RESTClient from "@/components/Client/RESTClient.js";
-import record from "@/store/record.js";
+import record from "@/store/recordData.js";
 import users from "@/store/users.js";
 import sinon from "sinon";
 import VueScrollTo from "vue-scrollto";
